backend/requirements.txt
# FastAPI Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-multipart==0.0.9

# Database
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin to 4.0.x for passlib compatibility

# Validation & Parsing
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0
PyYAML==6.0.1
python-json-logger==2.0.7

# Object Storage (S3/MinIO)
minio==7.2.3
boto3==1.34.34

# Background Tasks
celery==5.3.6
arq==0.25.0

# Utilities
python-dateutil==2.8.2
httpx==0.26.0
orjson==3.9.13
aiofiles==23.2.1

# Development
pytest>=8.2,<9.0
pytest-asyncio==0.24.0
pytest-cov==4.1.0

# Logging & Monitoring
structlog==24.1.0

